backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
redis
pydantic
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]